feat(home): rework homepage to split auth check from home page rendering
