+++ public/locales/nl/common.json
@@ -144,10 +144,10 @@
   "pages/index___geen-resultaten-voor-de-geselecteerde-filters": "Geen resultaten voor de geselecteerde filters.",
   "pages/index___hier-kom-je-er-alles-over-te-weten": "Hier kom je er alles over te weten.",
   "pages/index___laden": "Laden...",
-  "pages/index___plan-een-nieuw-bezoek-stap-fysiek-binnen-en-krijg-meteen-toegang-tot-het-digitale-archief-van-de-leeszaal-benieuwd-hoe-het-werkt": "Plan een nieuw bezoek, stap fysiek binnen en krijg meteen toegang tot het digitale archief van de leeszaal. Benieuwd hoe het werkt?",
+  "pages/index___logged-out-home-description": "Plan een nieuw bezoek, stap fysiek binnen en krijg meteen toegang tot het digitale archief van de leeszaal. Benieuwd hoe het werkt?",
   "pages/index___toon-alles-amount": "Toon alles ({{amount}})",
   "pages/index___vind-een-leeszaal": "Vind een leeszaal",
-  "pages/index___welkom-in-de-digitale-leeszaal": "Welkom in de digitale leeszaal",
+  "pages/index___logged-out-home-title": "Welkom in de digitale leeszaal",
   "pages/index___zoek": "Zoek",
   "pages/leeszaal/reading-room-slug/index___filters": "Filters",
   "pages/leeszaal/reading-room-slug/object-id/index___exporteer-metadata": "Exporteer metadata",
