+++ src/modules/home/components/LoggedInHome/__mocks__/VisitRequests.ts
@@ -1,4 +1,6 @@
-import { HeroProps, HeroRequest } from '../Hero.types';
+import { UserSchema } from '@auth/types';
+import { ReadingRoom } from '@shared/components';
+import { ComponentLink } from '@shared/types';
 
 export const heroTitle = 'Welkom in de digitale leeszaal';
 
@@ -11,6 +13,24 @@ export const heroLink = {
 };
 
 export const heroImage = { src: '/images/hero.png', alt: 'hero image' };
+
+export interface HeroProps {
+	title: string;
+	description: string;
+	link: ComponentLink;
+	image: HeroImage;
+	user?: UserSchema | null;
+	requests?: HeroRequest[];
+}
+
+export interface HeroImage {
+	src: string;
+	alt?: string;
+}
+
+export interface HeroRequest extends ReadingRoom {
+	status: 'access' | 'planned' | 'requested';
+}
 
 export const heroMock: HeroProps = {
 	title: heroTitle,
