+++ src/modules/shared/components/index.ts
@@ -1,7 +1,6 @@
 export * from './Blade';
 export * from './BladeManager';
 export * from './CardImage';
-export * from './Hero';
 export * from './Icon';
 export * from './ListNavigation';
 export * from './Loading';
